requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.7.7
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-Session==0.6.0
frozenlist==1.5.0
fsspec==2025.3.0
google-api-core==2.15.0
google-auth==2.28.1
google-cloud-core==2.4.3
google-cloud-firestore==2.15.0
google-cloud-run==0.10.0
google-cloud-secret-manager==2.19.0
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
multidict==6.2.0
numpy==1.24.3
openai==1.70.0
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
propcache==0.3.0
proto-plus==1.26.1
protobuf==4.25.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
tiktoken==0.5.2
tokenizers==0.21.1
tqdm==4.67.1
types-pytz==2025.1.0.20250204
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.18.3
